refactor: Put stubs under one file
