--- CSharp/src/BusinessApp.App.UnitTest/CommandStub.cs
@@ -1,13 +1,0 @@
-namespace BusinessApp.App.UnitTest
-{
-    public class CommandStub
-    {}
-
-    [Authorize]
-    public class AuthCommandStub : CommandStub
-    {}
-
-    [Authorize("Foo", "Bar")]
-    public class AuthRolesCommandStub : CommandStub
-    {}
-}
