--- CSharp/src/BusinessApp.App.UnitTest/QueryStub.cs
@@ -1,8 +1,0 @@
-namespace BusinessApp.App.UnitTest
-{
-    public class QueryStub : IQuery<ResponseStub>
-    {}
-
-    public class ResponseStub
-    {}
-}
